resolve similar param values in Url
