+++ dry_tests/models/content.py
@@ -1,0 +1,13 @@
+"""
+Content models
+"""
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class ContentValue:
+    """
+    Content Value Dataclass
+    """
+    value: str
+    count: int = None
